Push bot into a namespace and add tests
