+++ data.py
@@ -106,6 +106,11 @@ def preprocess_data(X, y, save=True, preset=None, subtract_mean=True):
     return X, y, n_samples, len(categories)
 
 
+def get_mean(dataset_path='data/data.hf5'):
+    h5f = h5py.File(dataset_path, 'r')
+    return h5f['mean'][:]
+
+
 def get_categories():
     '''Load categories names'''
