+++ deepanimebot/exceptions.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+
+class NotImage(Exception): pass
