+++ deepanimebot/shortcuts.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+import random
+
+
+def at_random(*messages):
+    return random.choice(messages)
