+++ deploy.py
@@ -20,9 +20,9 @@ import data
 Prediction = namedtuple('Prediction', 'rank category probability')
 
 
-def load_model(input_shape, n_outputs=100):
+def load_model(input_shape, n_outputs=100, model_name='model'):
     '''Loads and compiles pre-trained model to be used for real-time predictions'''
-    model = m.get_model(input_size=input_shape, n_outputs=n_outputs)
+    model = getattr(m, 'get_{}'.format(model_name))(input_size=input_shape, n_outputs=n_outputs)
     model.load_weights("pre_trained_weights/latest_model_weights.hdf5")
     model.compile(loss='categorical_crossentropy', optimizer='adam')
     return model
